requirements.txt
beautifulsoup4==4.12.2
google_api_python_client==2.89.0
google_auth_oauthlib==1.0.0
protobuf==4.25.1
psycopg2==2.9.9
python_dateutil==2.8.2
